fix(rbac): update legacy role metadata with help of file watcher
