+++ plugins/rbac-backend/src/file-permissions/csv-file-watcher.ts
@@ -22,6 +22,7 @@ import {
   RoleMetadataStorage,
 } from '../database/role-metadata';
 import {
+  mergeRoleMetadata,
   metadataStringToPolicy,
   policyToString,
   transformArrayToPolicy,
@@ -147,6 +148,35 @@ export class CSVFileWatcher {
         if (!(await tempEnforcer.hasPolicy(...policy))) {
           this.csvFilePolicies.removedPolicies.push(policy);
         }
+      }
+    }
+
+    // Check for policies that might need to be updated
+    // This will involve update "legacy" source in the role metadata if it exist in both the
+    // temp enforcer (csv file) and a role metadata storage.
+    // We will then add them back with the new source "csv-file"
+    const legacyRolesMetadata =
+      await this.roleMetadataStorage.filterRoleMetadata('legacy');
+
+    for (const legacyRole of legacyRolesMetadata) {
+      const legacyGroupPolicies = await tempEnforcer.getFilteredGroupingPolicy(
+        1,
+        legacyRole.roleEntityRef,
+      );
+      const legacyPolicies = await tempEnforcer.getFilteredPolicy(
+        0,
+        legacyRole.roleEntityRef,
+      );
+      if (legacyGroupPolicies.length > 0 || legacyPolicies.length > 0) {
+        const nonLegacyRole = mergeRoleMetadata(legacyRole, {
+          modifiedBy: CSV_PERMISSION_POLICY_FILE_AUTHOR,
+          source: 'csv-file',
+          roleEntityRef: legacyRole.roleEntityRef,
+        });
+        await this.roleMetadataStorage.updateRoleMetadata(
+          nonLegacyRole,
+          legacyRole.roleEntityRef,
+        );
       }
     }
